app/requirements.txt
# Core dependencies
fastapi>=0.115.12
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
psycopg2-binary>=2.9.6  # For PostgreSQL
alembic>=1.11.0  # Database migrations

# Production
gunicorn>=20.1.0
prometheus-client>=0.16.0
opentelemetry-api>=1.17.0
opentelemetry-sdk>=1.17.0
opentelemetry-exporter-otlp>=1.17.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2
pyyaml>=6.0
jinja2>=3.1.2

# For LLM service
requests>=2.30.0
numpy>=1.24.0

# Development
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

